santex-challenge-etl/requirements.txt
asn1crypto==1.5.1
asttokens==2.4.1
boto3==1.34.149
botocore==1.34.149
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
executing==2.0.1
filelock==3.15.4
fonttools==4.53.1
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
snowflake-connector-python==3.12.0
sortedcontainers==2.4.0
stack-data==0.6.3
threadpoolctl==3.5.0
tomlkit==0.13.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
